+++ yuno_sdk/android/src/main/kotlin/com/yuno_flutter/yuno_sdk_android/YunoSdkAndroidPlugin.kt
@@ -109,7 +109,8 @@ If you continue to have trouble, follow this discussion to get some support """,
                     call = call,
                     result = result,
                     context = context,
-                    activity = activity
+                    activity = activity,
+                    onState = this::onPaymentStateChange,
                 )
             }
             Key.enrollmentPayment -> {